1.lint性能优化（二）

+++ app/src/main/java/com/pictureair/photopass/editPhoto/filter/BlurFilter.java
@@ -43,9 +43,9 @@ public class BlurFilter extends Filter{
 	                        pixG = Color.green(pixColor); 
 	                        pixB = Color.blue(pixColor); 
 	                        
-	                        newR = newR + (int) (pixR * gauss[idx]); 
-	                        newG = newG + (int) (pixG * gauss[idx]); 
-	                        newB = newB + (int) (pixB * gauss[idx]); 
+	                        newR = newR + (pixR * gauss[idx]);
+	                        newG = newG + (pixG * gauss[idx]);
+	                        newB = newB + (pixB * gauss[idx]);
 	                        idx++; 
 	                    } 
 	                }
